Move encoders into libertem.io.dataset.mib
That way, we can make sure they are actually fully covered in the
tests, and we can more easily share them between different test modules.
